Lots of modifications made (a near full overhaul of the project)
This commit is nearly a complete overhaul of this repository.
Added channels : roles, questions-fréquentes, anciennes-vidéos, clash of code, dart | flutter
Changed the orders of the different channels to keep the same hierarchy since new channels were added (example the channel annonce was number two and became number 4 because of the channels roles and question) and changed their numbers in their filenames.
Added the new channels to the readme's
Modified some obsolete descriptions of channels (like the `autres` channel that doesn't exist anymore because it is now a forum)
Deleted some channels that don't exist/are not visible anymore